Move config back to /server. Refactor intl polyfill for React Native.
