+++ src/server/lib/errorHandler.js
@@ -1,4 +1,4 @@
-import config from '../../common/config';
+import config from '../config';
 
 export default function errorHandler(err, req, res, next) { // eslint-disable-line no-unused-vars
   const errorDetails = err.stack || err;
